minor clean ups and moving things around
- make the start, exchange, and finish wrappers do base64 for messages
  instead of byte arrays.
- move wasm stuff into its own directory.
- remove golang runtime file from repo (will add it back if we bundle
  the .wasm file too)
