+++ web/index.html
@@ -1,16 +1,7 @@
-<html>
-	<head>
-		<meta charset="utf-8"/>
-		<script src="wasm_exec.js"></script>
-		<script>
-			const go = new Go();
-			WebAssembly.instantiateStreaming(fetch("main.wasm"), go.importObject).then((result) => {
-				go.run(result.instance);
-			});
-		</script>
-	</head>
-	<body></body>
-
+<!doctype html>
+<meta charset=utf-8>
+<script src="wasm_exec.js"></script>
+<script src="wasm_exec.js"></script>
 <script>
 let dial = async (slot, config) => {
 	let pc = new RTCPeerConnection(config);
@@ -49,5 +40,14 @@ let dial = async (slot, config) => {
 	}
 	return dc
 }
+
+main = async () => {
+	window.cm = {};
+	const go = new Go();
+	let wasm = await WebAssembly.instantiateStreaming(fetch("main.wasm"), go.importObject);
+	go.run(wasm.instance);
+}
+
+main();
 </script>
-</html>
+<body>
